+++ internal/repositories/errors.go
@@ -1,0 +1,5 @@
+package repositories
+
+import "errors"
+
+var MaxSizeAchievedErr = errors.New("max size achieved")
