+++ internal/repositories/options.go
@@ -1,0 +1,13 @@
+package repositories
+
+type config struct {
+	maxLen int
+}
+
+type Option func(*config)
+
+func WithMaxSize(len int) Option {
+	return func(c *config) {
+		c.maxLen = len
+	}
+}
